+++ src/components/data-exporation/interactive/InteractiveExploration.tsx
@@ -25,7 +25,7 @@ const InteractiveExplorationPage: React.FC<{
             <Grid
                 item
                 style={{
-                    maxWidth: 800,
+                    maxWidth: 1100,
                     width: "inherit"
                 }}
             >
@@ -47,7 +47,7 @@ const InteractiveExplorationPage: React.FC<{
                 </Grid>
             </Grid>
             <Grid item>
-                <Card style={{ maxWidth: 800, width: "inherit" }}>
+                <Card style={{ maxWidth: 1100, width: "inherit" }}>
                     <CardHeader title="Overview" />
                     <CardContent className="markdown-body">
                         <Typography>
@@ -69,7 +69,7 @@ const InteractiveExplorationPage: React.FC<{
                 </Card>
             </Grid>
             <Grid item>
-                <Card style={{ maxWidth: 800, width: 800 }}>
+                <Card style={{ maxWidth: 1100, width: 1100 }}>
                     <CardHeader title="Direct Links" />
                     <CardContent className="markdown-body">
                         <Typography>