+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## 17 Jun 2022
+
+- `changed` width of data exploration cards
+
 ## 16 Jun 2022
 
 - `added` "Data Exploration" tab